Updating highlight function
Highlight function allows to check off items you already own to further personalize your shopping cart. ( Button is implemented and it works however I need to figure out how to pass key id into the function and also is compatible with state so that when pressed the only the pressed item is checked and not all items.